Add correct string recommendations [I18N-ACTION-1216]
Added, edited, and rearranged requirements. Added a link to the Design Principles doc.

These edits depend on additions to the glossary in a separate pull request. More work is needed here, particularly discussing when the choose `USVString` (which we don't mention) and making clear how to work with byte strings/encodings.