Add additional definitions for 'string' (w3c/i18n-actions#41)
There are multiple sources for the term 'string', including WEBIDL,
INFRA, Unicode, and XML Schema. In our discussions with RDF-star,
we have found that there is a need for generic string definitions
for use by JSON, Schema.org, RDF, and XML-based specs which are not
consistent with DOMString/USVString.

This PR needs to be accompanied by changes to the I18N-GLOSSARY in order
to be complete. This PR is for initial discussion purposes before we
embark on those changes.