Further (deep) rewrite
This change significantly reorganizes the section on strings in hopes
of making the definitions clearer and to reduce the overall
repetitive nature of the text.

It also moves the text about design-principles to a note at the top,
rather than burying it in the text.

The relationship of DOMString and USVString to surrogates is clarified
and pulled into one place.