Move the mustard ahead of the explanation
We have a general practice of putting the best practice text
before the explanation of the BP so that users linking into
specdev can read below the best practice rather than having
to scroll up and start the section over.

This change moves the string definition guidance ahead of all
of the text about string types.